# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bq — Birthday Queue",
    "version": "0.1.0"
  },
  "paths": {
    "/api/session": {
      "post": {
        "tags": [
          "convidado"
        ],
        "summary": "Create Session",
        "description": "Cria ou reidentifica o convidado. Idempotente por cookie: com cookie válido isto é\n`UPDATE` do MESMO convidado, nunca um segundo.",
        "operationId": "create_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "convidado"
        ],
        "summary": "Rename Session",
        "description": "RF-03. Renomeia **o mesmo** convidado: `UPDATE`, nunca `INSERT`.\n\n🔴 Se isto criasse um convidado novo, o cooldown de RF-09 morreria — e não por má fé: a\nprimeira pessoa que trocasse o apelido descobriria por acidente que o cooldown zerou, e\ncontaria para as outras. É a única defesa de cota que sobrou depois do corte de segurança,\ne ela cabe na escolha entre dois verbos SQL.",
        "operationId": "rename_session_api_session_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/suggestions/{suggestion_id}": {
      "delete": {
        "tags": [
          "convidado"
        ],
        "summary": "Remove Suggestion",
        "description": "RF-14. Só a própria, só enquanto não começou a tocar.\n\n**Não devolve a cota do cooldown**, e isso é a outra metade da mesma decisão: sem ela,\n\"sugerir e remover\" seria um jeito acidental de manter a fila inteira sob controle de uma\npessoa, e alguém descobre isso sem querer nos primeiros 20 minutos.",
        "operationId": "remove_suggestion_api_suggestions__suggestion_id__delete",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skip-votes": {
      "post": {
        "tags": [
          "convidado"
        ],
        "summary": "Cast Vote",
        "operationId": "cast_vote_api_skip_votes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "convidado"
        ],
        "summary": "Retract Vote",
        "operationId": "retract_vote_api_skip_votes_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/karaoke/start": {
      "post": {
        "tags": [
          "karaokê"
        ],
        "summary": "Start",
        "description": "A pessoa tocou INICIAR no próprio celular.\n\nA decisão inteira mora no maestro, sob o lock: checar aqui seria checar contra um estado que\npode mudar entre a leitura e o `_open`. Esta função só traduz a recusa para o contrato.",
        "operationId": "start_api_karaoke_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KaraokeStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KaraokeStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/host/session": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Host Session",
        "description": "RF-31. `secrets.compare_digest` não é paranoia de timing — é o jeito certo de comparar\nduas strings e custa o mesmo.",
        "operationId": "host_session_api_host_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Host Session Api Host Session Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/host/settings": {
      "get": {
        "tags": [
          "host"
        ],
        "summary": "Read Settings",
        "operationId": "read_settings_api_host_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsFull"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "host"
        ],
        "summary": "Patch Settings",
        "description": "RF-24. Efeito imediato, sem restart: grava em `setting`, recarrega o cache em memória e\nfaz broadcast — o /tv precisa passar a dizer `n de 4` na mesma hora.",
        "operationId": "patch_settings_api_host_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsFull"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/host/force-play": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Force Play",
        "description": "RF-26. É a saída manual do estado `idle` de ADR-005 — a rede que transforma \"silêncio\nquando a fila esvazia\" numa espera em vez de um beco.",
        "operationId": "force_play_api_host_force_play_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForcePlayIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForcePlayOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/host/suggestions/{suggestion_id}": {
      "delete": {
        "tags": [
          "host"
        ],
        "summary": "Host Remove",
        "description": "RF-29. O host remove qualquer sugestão, não só a própria.",
        "operationId": "host_remove_api_host_suggestions__suggestion_id__delete",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/host/karaoke/start": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Host Karaoke Start",
        "description": "O host começa a vez pela pessoa: o celular dela morreu, ou ela já está de pé na frente da\nTV com o microfone. Espelha o par `DELETE /api/suggestions/{id}` × o do host.",
        "operationId": "host_karaoke_start_api_host_karaoke_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KaraokeStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KaraokeStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/host/karaoke/cancel": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Host Karaoke Cancel",
        "description": "Encerra a vez em curso. `penalize=false` (o default) é o botão \"essa pessoa foi embora\":\nnão conta falta, porque quem decidiu foi o host e não a ausência dela.",
        "operationId": "host_karaoke_cancel_api_host_karaoke_cancel_post",
        "parameters": [
          {
            "name": "penalize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Penalize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Host Karaoke Cancel Api Host Karaoke Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/host/suggestions/{suggestion_id}/bump": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Host Bump",
        "description": "RF-30. Move uma sugestão para a frente da fila.\n\nNão altera cooldown nem cota de ninguém: é reordenação, não uma sugestão nova. E não fura o\nround-rank de quem já estava na frente por merecimento — apenas põe esta antes de todos, que\né o que o host pediu ao clicar.",
        "operationId": "host_bump_api_host_suggestions__suggestion_id__bump_post",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Host Bump Api Host Suggestions  Suggestion Id  Bump Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/host/suggestions/{suggestion_id}/last": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Host Send To Back",
        "description": "O par do bump: manda uma sugestão para o FIM da fila.\n\nNão é \"descer uma posição\" — ver o docstring de `queue.send_to_back`, onde está o porquê. Não\nchama `wake()`, ao contrário do bump: mandar para o fim nunca cria algo a tocar agora.",
        "operationId": "host_send_to_back_api_host_suggestions__suggestion_id__last_post",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Host Send To Back Api Host Suggestions  Suggestion Id  Last Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KaraokeStartIn": {
        "properties": {
          "suggestionId": {
            "type": "integer",
            "title": "Suggestionid"
          }
        },
        "type": "object",
        "required": [
          "suggestionId"
        ],
        "title": "KaraokeStartIn",
        "description": "🔴 O `suggestionId` no corpo não é decorativo: sem ele, um toque atrasado no botão do turno\nanterior iniciaria a vez de outra pessoa. Mesma função do `playId` no voto de skip."
      },
      "KaraokeStartOut": {
        "properties": {
          "playId": {
            "type": "integer",
            "title": "Playid"
          }
        },
        "type": "object",
        "required": [
          "playId"
        ],
        "title": "KaraokeStartOut"
      },
...
      "SessionIn": {
        "properties": {
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "SessionIn"
      },
      "SessionOut": {
        "properties": {
          "guestId": {
            "type": "integer",
            "title": "Guestid"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "cooldownUntilMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldownuntilms"
          }
        },
        "type": "object",
        "required": [
          "guestId",
          "nickname",
          "cooldownUntilMs"
        ],
        "title": "SessionOut"
      },
...
      "VoteIn": {
        "properties": {
          "playId": {
            "type": "integer",
            "title": "Playid"
          }
        },
        "type": "object",
        "required": [
          "playId"
        ],
        "title": "VoteIn"
      },
      "VoteOut": {
        "properties": {
          "votes": {
            "type": "integer",
            "title": "Votes"
          },
          "needed": {
            "type": "integer",
            "title": "Needed"
          },
          "youVoted": {
            "type": "boolean",
            "title": "Youvoted"
          }
        },
        "type": "object",
        "required": [
          "votes",
          "needed",
          "youVoted"
        ],
        "title": "VoteOut"
      },
...
    }
  }
}